// src/generated/library/LibraryEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Account } from "../live/Account";

export type LibraryEntry = { id: string, server: string, account: Account, championId: number, startedAt: number, duration: number, win: boolean, kills: number, deaths: number, assists: number, version: string, bookmarked: boolean, items: Array<number>, queueId: number, };